+++ cli/internal/dbmigrate/migrate_test.go
@@ -18,6 +18,7 @@ func TestMigrationApplyAndVerifyScriptsAreFailClosed(t *testing.T) {
 		"0017_device_key_recovery_rotation_0128",
 		"0018_device_trust_stabilization_01210",
 		"0019_minecraft_serverbridge_integrity_0135",
+		"0020_guard_migration_compatibility_stabilization_01310",
 		"database contains migrations unknown to this binary",
 		"UPDATE schema_migrations SET checksum=",
 		"pg_advisory_lock(718033100100)",